Update sq.csv
Some of the translations are obtained from google translate and are not those we use in our statistical dictionary . There are no direct translations for Jitter points, but a clear explanation can be added if necessary. 
Thank you for the opportunity to contribute and make things easier for users. I teach R at university and Esquisse is the package that I start in parallel with ggplot , it makes things easier for my students. 
Thank you and keep going!